requirements.txt
langchain>=0.1.0
langchain-core>=0.1.0
langgraph>=0.0.15
langchain-groq>=0.0.3
langchain-community>=0.0.13
gradio>=4.14.0
requests>=2.31.0
transformers>=4.37.2
torch>=2.1.2
chromadb>=0.4.22
datasets>=2.16.1
duckduckgo-search>=4.1.1
fastembed>=0.1.1
onnxruntime>=1.16.3
python-dotenv>=1.0.0
pandas>=2.1.0
numpy>=1.24.0
Pillow>=10.0.0
kagglehub>=0.1.0
backoff>=2.2.1
pydantic>=2.5.0 